Create server directory
